Added a web interface to netplot.
The netplot_server.py is a server that receives the netplot
commands from a client and saves the plots as json files.

Once the json files have been created a webserver can use
this location as the web server path root. If a broweser
is then pointed at this location then the index.html will
display the plots.

The legends.html can also be displayed to check which plot ID
against the entire plot title text.